Add settings database

+++ astrology/src/main/java/com/android/pam/astrology/domain/repository/IAstrologySettingsRepository.kt
@@ -7,5 +7,5 @@ interface IAstrologySettingsRepository {
     fun setSettings(settings: Settings)
     fun getSettings(): Settings
     fun getLocation(): Location
-    fun getRefreshRate(): Int
+    fun getRefreshRate(): Double
 }
